Fix Larva Build Queue
Fixes issue where a Larva's build queue can have hanging items causing problems with morph commands on the larva or Hydralisks and Mutalisks by having a morph command and stop command issued on the same frame. This ensures the build_queue is clear before attempting to add the desired unit type to the build_queue.

Directly erasing maintains the 1.16.1 behavior of the resources still being consumed by this situation, while keeping the larva and it's future unit type functioning properly as in 1.16.1.