Special Case Lurker Devourer And Guardian
This removes the loop I made to clear build queues for larva - this deviates from BW behavior in edge cases since in 1.16 and Remaster you can brick a larva with morph/stop on same frame 5 times, and previous fix prevented that.

I observed that 1.16 puts the Lurker/Guardian/Devourer types at the front of the build queue on the unit instead of the back. In OpenBW there is no function to do this with a static_vector, so if it pushes to the end successfully I check if it is a Lurker/Guardian/Devourer morph, then push indexes 0 through 3 down 1, then puts the correct unit type (what was pushed to the end) at the front of the build_queue. Behaviorally this is what BW does, but I'm not sure it does it exactly this way.

I feel this would be cleaner if there was a way to push to the front of the build queue, but that doesn't exist for a static_vector from what I can see.